Overhaul warning
It is starting to look more and more like I will need to overhaul the UI system to enable for the system to work as it is supposed to, this will include changing how I am handling the UI components. I will most likely turn to an OOP style as I can simply implement it then without much effort and then later on swap the layout of it.